# Improve performance of PacketBuffer class by using a List<>.