As said in the issue #1, I wasn't too happy of the runtimes of `ExtractFirstValidPacket()`, especially after implementing the FileProvider, and seeing how slow packets were extracted from the buffer.
To try and compare each method runtime, I benchmarked them to see how drastic the difference actually was:
| Method         | Mean     | Error   | StdDev  | Ratio | Gen0   | Allocated | Alloc Ratio |
|--------------- |---------:|--------:|--------:|------:|-------:|----------:|------------:|
| ExtractAll_Old | 667.9 ns | 4.96 ns | 4.14 ns |  1.00 | 0.0391 |     656 B |        1.00 |
| ExtractAll_New | 129.4 ns | 1.18 ns | 1.05 ns |  0.19 | 0.0143 |     240 B |        0.37 |

And came down to a staggering 5x improved runtime and less than half of the memory used! I think the previous method, by also analysing the buffer with Array.IndexOf(), turned out to be very slow.
Closes #1