@mslago 2025-07-05
I should also add checking of a packet even if there is no termination byte. 
Say the first termination byte that appears is on index 17 (bigger than one packet). We also need to check if everything that comes before that is a valid packet.