@mslago 2025-07-06
Okay! After intense testing, everything seems exactly like it was with the previous method. I'll merge.